Fix parameter type error in get_ber()

+++ dt_c_api.py
@@ -49,8 +49,8 @@ def get_ber(iamp, qamp):
     """ Calculate number of total and erroneously decoded symbols
     """
     c_size = c_int(len(iamp))
-    c_iamp = (c_double*len(iamp))(*iamp)
-    c_qamp = (c_double*len(iamp))(*qamp)
+    c_iamp = (c_int*len(iamp))(*iamp)
+    c_qamp = (c_int*len(iamp))(*qamp)
     c_numerr = c_int(0)
     c_numbit = c_int(0)
     rc = _libdmr.bercalc(c_iamp, c_qamp, c_size, byref(c_numerr), byref(c_numbit))
